Refactor : "수정 전체적인 연관관계 매핑"
