+++ src/main/java/com/sparta/devstar_be/share/Share.java
@@ -31,8 +31,12 @@ public class Share {
     private String name;
     private String major;
 
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private User user;
+
     @OneToMany(mappedBy = "share")
-    private Set<Like> shareLikes = new HashSet<>();
+    private List<Like> shareLikeList = new ArrayList<>();
 
     @Column(name = "like_count")
     private int likeCount;
@@ -71,7 +75,5 @@ public class Share {
 //        this.major = user.getMajor();
 //    }
   
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
+
 }
